- order schemas liste - add check pdf_vision - add info formulaires

+++ app/controllers/schemas_controller.rb
@@ -7,7 +7,7 @@ class SchemasController < ApplicationController
   def index
     # récupérer la liste des schémas triés par ordre croissant
     @schemas = current_user.statut == 'DCB' ? current_user.schemas.joins(:gestion_schemas).distinct : Schema.where(statut: 'valide')
-    @schemas = @schemas.order(created_at: :desc)
+    @schemas = @schemas.order(updated_at: :desc)
     # recherche filtres ransack
     @q = @schemas.ransack(params[:q])
     @schemas = @q.result.includes(:programme, :user)
@@ -57,22 +57,31 @@ class SchemasController < ApplicationController
     respond_to do |format|
       format.html
       format.pdf do
-        url = pdf_vision_schema_url(@schema)
-        tmp_path = UrlToPdfJob.perform_now(url)
-        # envoyer le pdf en réponse
-        pdf_data = File.read(tmp_path)
-        send_data(pdf_data,
-                  filename: "schema_P#{@schema.programme.numero}.pdf",
-                  type: "application/pdf",
-                  disposition: "attachment") # inline open in browser
-        # disposition: "attachment") # default # download
-        # ouvrir le fichier temporaire et attacher au modèle Schema
-        File.open(tmp_path) do |file|
-          @schema.document_pdf.attach(io: file, filename: "schema_P#{@schema.programme.numero}.pdf")
-        end
+        if @schema.document_pdf.attached? # vérifiez si un pdf est attaché
+          send_data(
+            @schema.document_pdf.download, # envoyez le pdf attaché
+            filename: @schema.document_pdf.filename.to_s,
+            type: "application/pdf",
+            disposition: "attachment"
+          )
+        else
+          url = pdf_vision_schema_url(@schema)
+          tmp_path = UrlToPdfJob.perform_now(url)
+          # envoyer le pdf en réponse
+          pdf_data = File.read(tmp_path)
+          send_data(pdf_data,
+                    filename: "schema_P#{@schema.programme.numero}.pdf",
+                    type: "application/pdf",
+                    disposition: "attachment") # inline open in browser
+          # disposition: "attachment") # default # download
+          # ouvrir le fichier temporaire et attacher au modèle Schema
+          File.open(tmp_path) do |file|
+            @schema.document_pdf.attach(io: file, filename: "schema_P#{@schema.programme.numero}.pdf")
+          end
 
-        # assurez-vous de supprimer le fichier temporaire après son utilisation
-        File.delete(tmp_path)
+          # assurez-vous de supprimer le fichier temporaire après son utilisation
+          File.delete(tmp_path)
+        end
       end
     end
   end
